backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
langchain
langchain-google-genai
google-generativeai
SQLAlchemy
fastapi-cors
langchain-community